Change where functions are defined
