+++ .gitignore
@@ -11,7 +11,8 @@ node_modules
 !.env.example
 .cloudflare
 robots.txt
-src\routes\[lang=lang]\(manual)\manual\workspace\+page.svelte
+admin/
+signup/
 changelog.txt
 vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
